+++ sdk/src/main/java/com/ciscowebex/androidsdk/phone/Call.java
@@ -22,18 +22,17 @@
 
 package com.ciscowebex.androidsdk.phone;
 
-import java.util.List;
-import java.util.Set;
-
+import android.app.Notification;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-
+import android.util.Pair;
 import android.util.Size;
+import android.view.View;
 
 import com.ciscowebex.androidsdk.CompletionHandler;
 
-import android.view.View;
-import android.util.Pair;
+import java.util.List;
+import java.util.Set;
 
 /**
  * A Call represents a media call on Cisco Webex.
@@ -241,6 +240,7 @@ public interface Call {
     /**
      * Specify how the remote video adjusts its content to be render in a view.
      *
+     * @param mode the remote video render mode
      * @since 2.6.0
      */
     void setRemoteVideoRenderMode(VideoRenderMode mode);
@@ -248,9 +248,28 @@ public interface Call {
     /**
      * Specify the video layout for the active speaker and other attendees in the group video meeting.
      *
+     * @param layout the video layout mode
      * @since 2.6.0
+     * @deprecated please use {@link Call#setCompositedVideoLayout(MediaOption.CompositedVideoLayout)}
      */
-    void setVideoLayout(MediaOption.VideoLayout layout);
+    void setVideoLayout(MediaOption.CompositedVideoLayout layout);
+
+    /**
+     * Specify the composited video layout for the active speaker and other attendees in the group video meeting.
+     *
+     * @param layout the composited video layout mode
+     * @since 2.8.0
+     */
+    void setCompositedVideoLayout(MediaOption.CompositedVideoLayout layout);
+
+    /**
+     * Specify the composited video layout for the active speaker and other attendees in the group video meeting.
+     *
+     * @param layout   the composited video layout mode
+     * @param callback A closure to be executed when completed, with error if the invocation is illegal or failed, otherwise nil.
+     * @since 2.8.0
+     */
+    void setCompositedVideoLayout(MediaOption.CompositedVideoLayout layout, @Nullable CompletionHandler<Void> callback);
 
     /**
      * @return The local video render view dimensions (points) of this call.
@@ -436,10 +455,22 @@ public interface Call {
 
     /**
      * Start content sharing.
-     *
+     * Invalid if targetSdkVersion is 29 or higher, please use {@link Call#startSharing(Notification, int, CompletionHandler)}
      * @since 1.4
+     * @deprecated
      */
     void startSharing(@NonNull CompletionHandler<Void> callback);
+
+    /**
+     * Start content sharing. The notification and it's id are only work for targetSdkVersion is 29 or higher.
+     *
+     * @param notification   The foreground notification when sharing, take effect if targetSdkVersion is 29 or higher.
+     * @param notificationId The id of this notification, take effect if targetSdkVersion is 29 or higher, must a positive number.
+     * @param callback       A closure to be executed when completed, with error if the invocation is illegal or failed, otherwise nil.
+     * @see <a href="https://developer.android.com/reference/android/media/projection/MediaProjectionManager#getMediaProjection(int,%20android.content.Intent)">Google Developers Page</a>
+     * @since 2.8.0
+     */
+    void startSharing(@Nullable Notification notification, int notificationId, @NonNull CompletionHandler<Void> callback);
 
     /**
      * Stop content sharing.