+++ sdk/src/main/java/com/ciscowebex/androidsdk/phone/CallMembership.java
@@ -100,10 +100,12 @@ public interface CallMembership {
     State getState();
 
     /**
-     * @return The email address of the person in this CallMembership.
-     * @since 0.1
+     * Notice, removed the getEmail() function, due to the email address can no longer get, since privacy protection reason.
+     * Please use this function to show a display name instead.
+     * @return The display name of the person in this CallMembership.
+     * @since 2.8.0
      */
-    String getEmail();
+    String getDisplayName();
 
     /**
      * @return The SIP address of the person in this CallMembership.